+++ clientcontroller/babylon/babylon.go
@@ -5,18 +5,19 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/babylonlabs-io/babylon/client/babylonclient"
+	"github.com/babylonlabs-io/babylon/v3/client/babylonclient"
 
 	"github.com/babylonlabs-io/finality-provider/finality-provider/proto"
+	"github.com/babylonlabs-io/finality-provider/finality-provider/signingcontext"
 
 	sdkErr "cosmossdk.io/errors"
 	sdkmath "cosmossdk.io/math"
-	bbnclient "github.com/babylonlabs-io/babylon/client/client"
-	bbntypes "github.com/babylonlabs-io/babylon/types"
-	btcctypes "github.com/babylonlabs-io/babylon/x/btccheckpoint/types"
-	btclctypes "github.com/babylonlabs-io/babylon/x/btclightclient/types"
-	btcstakingtypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
-	bsctypes "github.com/babylonlabs-io/babylon/x/btcstkconsumer/types"
+	bbnclient "github.com/babylonlabs-io/babylon/v3/client/client"
+	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
+	btcctypes "github.com/babylonlabs-io/babylon/v3/x/btccheckpoint/types"
+	btclctypes "github.com/babylonlabs-io/babylon/v3/x/btclightclient/types"
+	btcstakingtypes "github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
+	bsctypes "github.com/babylonlabs-io/babylon/v3/x/btcstkconsumer/types"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/schnorr"
 	"github.com/btcsuite/btcd/btcutil"
@@ -112,6 +113,10 @@ func (bc *BabylonController) reliablySendMsgs(msgs []sdk.Msg, expectedErrs []*sd
 	)
 }
 
+func (bc *BabylonController) GetFpPopContextV0() string {
+	return signingcontext.FpPopContextV0(bc.cfg.ChainID, signingcontext.AccBTCStaking.String())
+}
+
 // RegisterFinalityProvider registers a finality provider via a MsgCreateFinalityProvider to Babylon
 // it returns tx hash and error
 // If chainID is empty, then it means the FP is a Babylon FP
@@ -139,7 +144,7 @@ func (bc *BabylonController) RegisterFinalityProvider(
 		Pop:         &bbnPop,
 		Commission:  commission,
 		Description: &sdkDescription,
-		ConsumerId:  chainID,
+		BsnId:       chainID,
 	}
 
 	res, err := bc.reliablySendMsg(msg, emptyErrs, emptyErrs)
@@ -304,7 +309,8 @@ func (bc *BabylonController) QueryFinalityProviders() ([]*btcstakingtypes.Finali
 	}
 
 	for {
-		res, err := bc.bbnClient.QueryClient.FinalityProviders(pagination)
+		// NOTE: empty BSN ID means querying all Babylon finality providers
+		res, err := bc.bbnClient.QueryClient.FinalityProviders("", pagination)
 		if err != nil {
 			return nil, fmt.Errorf("failed to query finality providers: %w", err)
 		}
@@ -319,14 +325,14 @@ func (bc *BabylonController) QueryFinalityProviders() ([]*btcstakingtypes.Finali
 	return fps, nil
 }
 
-func (bc *BabylonController) QueryConsumerFinalityProviders(consumerID string) ([]*bsctypes.FinalityProviderResponse, error) {
-	var fps []*bsctypes.FinalityProviderResponse
+func (bc *BabylonController) QueryConsumerFinalityProviders(bsnID string) ([]*btcstakingtypes.FinalityProviderResponse, error) {
+	var fps []*btcstakingtypes.FinalityProviderResponse
 	pagination := &sdkquery.PageRequest{
 		Limit: 100,
 	}
 
 	for {
-		res, err := bc.bbnClient.QueryClient.QueryConsumerFinalityProviders(consumerID, pagination)
+		res, err := bc.bbnClient.QueryClient.FinalityProviders(bsnID, pagination)
 		if err != nil {
 			return nil, fmt.Errorf("failed to query finality providers: %w", err)
 		}
@@ -469,11 +475,11 @@ func (bc *BabylonController) InsertSpvProofs(submitter string, proofs []*btcctyp
 // RegisterConsumerChain registers a consumer chain via a MsgRegisterChain to Babylon
 func (bc *BabylonController) RegisterConsumerChain(id, name, description, ethL2FinalityContractAddress string) (*types.TxResponse, error) {
 	msg := &bsctypes.MsgRegisterConsumer{
-		Signer:                       bc.MustGetTxSigner(),
-		ConsumerId:                   id,
-		ConsumerName:                 name,
-		ConsumerDescription:          description,
-		EthL2FinalityContractAddress: ethL2FinalityContractAddress,
+		Signer:                        bc.MustGetTxSigner(),
+		ConsumerId:                    id,
+		ConsumerName:                  name,
+		ConsumerDescription:           description,
+		RollupFinalityContractAddress: ethL2FinalityContractAddress,
 	}
 
 	res, err := bc.reliablySendMsg(msg, emptyErrs, emptyErrs)