+++ clientcontroller/babylon/consumer.go
@@ -6,11 +6,11 @@ import (
 	"strings"
 
 	sdkErr "cosmossdk.io/errors"
-	"github.com/babylonlabs-io/babylon/client/babylonclient"
-	bbnclient "github.com/babylonlabs-io/babylon/client/client"
-	bbntypes "github.com/babylonlabs-io/babylon/types"
-	btcstakingtypes "github.com/babylonlabs-io/babylon/x/btcstaking/types"
-	finalitytypes "github.com/babylonlabs-io/babylon/x/finality/types"
+	"github.com/babylonlabs-io/babylon/v3/client/babylonclient"
+	bbnclient "github.com/babylonlabs-io/babylon/v3/client/client"
+	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
+	btcstakingtypes "github.com/babylonlabs-io/babylon/v3/x/btcstaking/types"
+	finalitytypes "github.com/babylonlabs-io/babylon/v3/x/finality/types"
 	"github.com/btcsuite/btcd/btcec/v2"
 	"github.com/btcsuite/btcd/btcec/v2/schnorr"
 	cmtcrypto "github.com/cometbft/cometbft/proto/tendermint/crypto"
@@ -20,6 +20,7 @@ import (
 
 	"github.com/babylonlabs-io/finality-provider/clientcontroller/api"
 	fpcfg "github.com/babylonlabs-io/finality-provider/finality-provider/config"
+	"github.com/babylonlabs-io/finality-provider/finality-provider/signingcontext"
 	"github.com/babylonlabs-io/finality-provider/types"
 )
 
@@ -96,6 +97,14 @@ func (bc *BabylonConsumerController) reliablySendMsgs(msgs []sdk.Msg, expectedEr
 		expectedErrs,
 		unrecoverableErrs,
 	)
+}
+
+func (bc *BabylonConsumerController) GetFpRandCommitContext() string {
+	return signingcontext.FpRandCommitContextV0(bc.cfg.ChainID, signingcontext.AccFinality.String())
+}
+
+func (bc *BabylonConsumerController) GetFpFinVoteContext() string {
+	return signingcontext.FpFinVoteContextV0(bc.cfg.ChainID, signingcontext.AccFinality.String())
 }
 
 // CommitPubRandList commits a list of Schnorr public randomness via a MsgCommitPubRand to Babylon