+++ finality-provider/cmd/fpd/daemon/recover_proof.go
@@ -6,7 +6,7 @@ import (
 	"math"
 	"path/filepath"
 
-	bbntypes "github.com/babylonlabs-io/babylon/types"
+	bbntypes "github.com/babylonlabs-io/babylon/v3/types"
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/spf13/cobra"