+++ types/txresponse.go
@@ -1,7 +1,7 @@
 package types
 
 import (
-	"github.com/babylonlabs-io/babylon/client/babylonclient"
+	"github.com/babylonlabs-io/babylon/v3/client/babylonclient"
 )
 
 // TxResponse handles the transaction response in the interface ConsumerController